`Labels.split` and `Labels.make_training_splits` (#98)
* Add "all" embed target, clean up embedding docs, dedup, tests

* Video typing things and cleanup

* Labels.split

* Add provenance (filename) at load time

* Provenance serialization

* Add provenance to Labels.split

* Add Labels.make_training_splits

* Update docs

* Bump version

* Test fixes

+++ sleap_io/io/video.py
@@ -538,7 +538,7 @@ class HDF5Video(VideoBackend):
                 self.image_format = ds.attrs["format"]
 
             if "frame_numbers" in ds.parent:
-                frame_numbers = ds.parent["frame_numbers"][:]
+                frame_numbers = ds.parent["frame_numbers"][:].astype(int)
                 self.frame_map = {frame: idx for idx, frame in enumerate(frame_numbers)}
                 self.source_inds = frame_numbers
